ADDED: search params +skip +limit +favorite
