+++ controllers/auth/getCurrent.js
@@ -1,0 +1,7 @@
+const getCurrent = async (req, res) => {
+  const { email, subscription } = req.user;
+
+  res.json({ email, subscription });
+};
+
+module.exports = getCurrent;
