+++ controllers/auth/index.js
@@ -1,0 +1,12 @@
+const { ctrlWrapper } = require('../../helpers');
+const register = require('./register');
+const login = require('./login');
+const getCurrent = require('./getCurrent');
+const logout = require('./logout');
+
+module.exports = {
+  register: ctrlWrapper(register),
+  login: ctrlWrapper(login),
+  getCurrent: ctrlWrapper(getCurrent),
+  logout: ctrlWrapper(logout),
+};
