+++ routes/api/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const { validateBody, authenticate } = require('../../middlewares');
 const { schemas } = require('../../models/user');
-const ctrl = require('../../controllers/auth/auth');
+const ctrl = require('../../controllers/auth');
 
 const router = express.Router();
